+++ src/extension.ts
@@ -1,8 +1,30 @@
-import * as path from 'path';
 import * as vscode from 'vscode';
 
+const appName = 'TriggerMate';
+const cfgName = `${appName
+  .substring(0, 1)
+  .toLocaleLowerCase()}${appName.substring(1)}`;
+const cfgProp = 'fileTriggers';
+
 const workspacePath = vscode.workspace.workspaceFolders?.[0].uri.fsPath;
-const outputChannel = vscode.window.createOutputChannel('TriggerMate');
+const outputChannel = vscode.window.createOutputChannel(appName);
+
+type FileTriggerCommand = 'terminal' | 'task';
+
+interface FileTrigger {
+  file: string;
+  command?: string;
+  autoExecute?: boolean;
+  commandType?: FileTriggerCommand;
+
+  ignoreCreateEvents?: boolean;
+  ignoreChangeEvents?: boolean;
+  ignoreDeleteEvents?: boolean;
+
+  commandDebounce?: number;
+  commandLastTrigger?: number;
+}
+
 let watchers: vscode.FileSystemWatcher[] = [];
 
 export function activate(context: vscode.ExtensionContext) {
@@ -11,40 +33,71 @@ export function activate(context: vscode.ExtensionContext) {
     return;
   }
 
-  const updateWatchers = () => {
-    // Dispose of existing watchers
-    watchers.forEach((watcher) => watcher.dispose());
+  const resetWatchers = (): void => {
+    watchers.forEach((watcher: vscode.FileSystemWatcher) => watcher.dispose());
     watchers = [];
+  };
+  const loadTriggers = (): FileTrigger[] | undefined => {
+    const config = vscode.workspace.getConfiguration(cfgName);
 
-    const config = vscode.workspace.getConfiguration('triggerMate');
-    const fileTriggers = config.get<{ file: string; command: string }[]>(
-      'fileTriggers',
-      []
-    );
-
-    if (!Array.isArray(fileTriggers)) {
-      outputChannel.appendLine('Invalid configuration for fileTriggers.');
-      return;
+    const loadedFileTriggers = config.get<FileTrigger[]>(cfgProp, []);
+    if (!Array.isArray(loadedFileTriggers)) {
+      outputChannel.appendLine('FileTriggers must be an array.');
+      return undefined;
     }
 
-    fileTriggers.forEach((entry) => {
-      if (!entry.file) {
+    return loadedFileTriggers;
+  };
+
+  const updateWatchers = () => {
+    // Dispose of existing watchers
+    resetWatchers();
+
+    const fileTriggers = loadTriggers();
+    fileTriggers?.forEach((fileTrigger: FileTrigger) => {
+      const {
+        file,
+        ignoreChangeEvents,
+        ignoreCreateEvents,
+        ignoreDeleteEvents,
+      } = fileTrigger;
+      if (!file) {
         outputChannel.appendLine(
-          'Invalid entry in fileTriggers configuration.'
+          'Invalid entry in fileTriggers configuration: file property required.'
         );
         return;
       }
 
-      const filePath = path.join(workspacePath, entry.file);
-
-      const watcher = vscode.workspace.createFileSystemWatcher(filePath);
-      watcher.onDidChange(() => handleFileChange(entry));
-      watcher.onDidCreate(() => handleFileChange(entry));
-      watcher.onDidDelete(() => handleFileChange(entry));
+      const pattern = new vscode.RelativePattern(workspacePath, file);
+      const watcher: vscode.FileSystemWatcher =
+        vscode.workspace.createFileSystemWatcher(
+          pattern,
+          ignoreCreateEvents,
+          ignoreChangeEvents,
+          ignoreDeleteEvents
+        );
+      watcher.onDidChange((uri: vscode.Uri) =>
+        handleFileChange(fileTrigger, uri, 'onChange')
+      );
+      watcher.onDidCreate((uri: vscode.Uri) =>
+        handleFileChange(fileTrigger, uri, 'onCreate')
+      );
+      watcher.onDidDelete((uri: vscode.Uri) =>
+        handleFileChange(fileTrigger, uri, 'onDelete')
+      );
 
       context.subscriptions.push(watcher);
       watchers.push(watcher);
-      outputChannel.appendLine(`Watching ${filePath}...`);
+
+      outputChannel.appendLine(
+        `Watching [${pattern.baseUri}, ${pattern.pattern}] - (${[
+          !ignoreCreateEvents ? 'create' : '',
+          !ignoreChangeEvents ? 'change' : '',
+          !ignoreDeleteEvents ? 'delete' : '',
+        ]
+          .filter(Boolean)
+          .join(', ')})`
+      );
     });
   };
 
@@ -53,70 +106,115 @@ export function activate(context: vscode.ExtensionContext) {
 
   // Update watchers when configuration changes
   context.subscriptions.push(
-    vscode.workspace.onDidChangeConfiguration((e) => {
-      if (e.affectsConfiguration('triggerMate.fileTriggers')) {
-        updateWatchers();
+    vscode.workspace.onDidChangeConfiguration(
+      (e: vscode.ConfigurationChangeEvent) => {
+        if (e.affectsConfiguration(`${cfgName}.${cfgProp}`)) {
+          updateWatchers();
+        }
       }
-    })
+    )
   );
 
-  outputChannel.appendLine('TriggerMate extension activated.');
+  outputChannel.appendLine(`${appName} extension activated.`);
 }
 
-function handleFileChange(entry: {
-  file: string;
-  command?: string;
-  autoExecute?: boolean;
-}) {
+function handleFileChange(
+  fileTrigger: FileTrigger,
+  uri: vscode.Uri,
+  eventName: string
+) {
+  const {
+    file,
+    autoExecute,
+    command,
+    commandType,
+    commandDebounce,
+    commandLastTrigger,
+  } = fileTrigger;
+
+  if (commandDebounce) {
+    if (commandLastTrigger) {
+      if (new Date().getTime() - commandLastTrigger < commandDebounce) {
+        return;
+      }
+    }
+
+    fileTrigger.commandLastTrigger = new Date().getTime();
+  }
+
   if (!workspacePath) {
     outputChannel.appendLine('No workspace path available.');
     return;
   }
-
-  if (!entry.command) {
-    vscode.window.showInformationMessage(`${entry.file} has changed.`);
+  if (!command) {
+    vscode.window.showInformationMessage(`${file} has changed.`);
     return;
   }
 
-  if (entry.autoExecute) {
+  if (autoExecute) {
     vscode.window.showInformationMessage(
-      `${entry.file} has changed. Running \`${entry.command}\`.`
+      `${file} has changed. Running \`${command}\`.`
     );
-    runCommand(entry.command, workspacePath);
+    if (commandType === 'task') {
+      runTask(fileTrigger, uri, eventName);
+    } else {
+      runCommand(fileTrigger, uri, eventName);
+    }
   } else {
     vscode.window
       .showInformationMessage(
-        `${entry.file} has changed. Run \`${entry.command}\`?`,
+        `${file} has changed. Run \`${command}\`?`,
         'Yes',
         'No'
       )
       .then((selection) => {
         if (selection === 'Yes') {
-          if (!entry.command) {
-            vscode.window.showErrorMessage(
-              'No command specified for this file trigger.'
-            );
-            return;
+          if (commandType === 'task') {
+            runTask(fileTrigger, uri, eventName);
+          } else {
+            runCommand(fileTrigger, uri, eventName);
           }
-          runCommand(entry.command, workspacePath);
         }
       });
   }
 }
 
-function runCommand(command: string, workspacePath: string) {
+function runCommand(
+  fileTrigger: FileTrigger,
+  uri: vscode.Uri,
+  eventName: string
+) {
+  const { command, file } = fileTrigger;
+
   try {
-    const terminal = vscode.window.createTerminal('TriggerMate');
+    const terminal = vscode.window.createTerminal(appName);
     terminal.sendText(`cd ${workspacePath} && ${command}`);
     terminal.show();
-    outputChannel.appendLine(`Running command: ${command}`);
+    outputChannel.appendLine(
+      `Running command: ${command} for ${file} at ${uri} on ${eventName}`
+    );
   } catch (error) {
     outputChannel.appendLine(`Error running command: ${error}`);
   }
 }
 
+function runTask(fileTrigger: FileTrigger, uri: vscode.Uri, eventName: string) {
+  const { command, file } = fileTrigger;
+
+  vscode.tasks.fetchTasks().then((availableTasks: vscode.Task[]) => {
+    availableTasks.forEach((task: vscode.Task) => {
+      if (task.name === command) {
+        outputChannel.appendLine(
+          `Running task: ${command} for ${file} at ${uri} on ${eventName}`
+        );
+        vscode.tasks.executeTask(task);
+      }
+    });
+  });
+}
+
 export function deactivate() {
-  outputChannel.appendLine('TriggerMate extension deactivated.');
+  outputChannel.appendLine(`${appName} extension deactivated.`);
   // Dispose of all watchers
   watchers.forEach((watcher) => watcher.dispose());
 }